pyproject.toml
[tool.poetry]
name = "sas-as-tool-examples"
version = "0.1.0"
description = "An example of using SAS with saspy as a tool for langchain and langgraph."
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
saspy = "^5.100.2"
notebook = "^7.2.2"
langgraph = "^0.2.18"
langchain_core = "^0.2.38"
langchain_community = "^0.2.16"
langchain_openai = "^0.1.23"
langchain_ollama = "^0.1.3"
langchain_experimental = "^0.0.65"
pandas = "^2.2.2"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
